refactor&chore: split root page.tsx and disable eslint rule
