+++ src/app/page.tsx
@@ -1,26 +1,10 @@
 import {useBlockingQR, BlockingQR} from '@/app/blocking-qr/dialog';
-import {useUserAccessHashes} from '@/components/useraccesshashes-provider';
-import {UserDetails} from '@/types/user-details';
 import {useAppContext, useAppContextRef} from '@/app.context';
-import {FeedItem} from '@/network/friendly-client';
 import {useEffect, useMemo, useState, useCallback} from 'react';
-import {toast} from 'sonner';
 import {Avatar, AvatarFallback, AvatarImage} from '@/components/ui/avatar';
 import {Badge} from '@/components/ui/badge';
 import {Separator} from '@/components/ui/separator';
-import QRCode from 'react-qr-code';
-import {
-    Activity,
-    Check,
-    Copy,
-    Heart,
-    Loader2,
-    LogOut,
-    Pencil,
-    QrCodeIcon,
-    X,
-    RotateCcw,
-} from 'lucide-react';
+import {Activity, Loader2, LogOut, Pencil} from 'lucide-react';
 import {Button} from '@/components/ui/button';
 import {useNavigate} from 'react-router';
 import {useBackend} from '@/backend.context';
@@ -28,455 +12,16 @@ import {formatNetworkError} from '@/services/backend-service';
 import {
     createFileLink,
     createFriendInviteLink,
-    truncateString,
+    getAvatarFallbackForNickname,
 } from '@/lib/utils';
-import {useQuery, useQueryClient} from '@tanstack/react-query';
+import {useQuery} from '@tanstack/react-query';
 import {useSession} from '@/components/session-provider';
 import {useTranslations} from 'use-intl';
 import {EditProfileDialog} from '@/app/edit/dialog';
-import * as Dialog from '@radix-ui/react-dialog';
 import {ProfileDescription} from '@/components/profile-description';
-import {FriendsBlock} from './friends/friends-block';
-import {StyledDialogWrapper} from '@/components/styled-dialog-wrapper';
-
-type SwipeDirection = 'left' | 'right';
-
-function getFeedItemKey(item: FeedItem) {
-    return `${item.details.id}-${item.isRequest ? 'request' : 'suggested'}`;
-}
-
-function FeedEmptyState() {
-    const t = useTranslations('profile.feed');
-
-    return (
-        <div className="flex min-h-64 flex-col items-center justify-center rounded-xl border border-dashed border-zinc-200 bg-zinc-50 px-6 text-center dark:border-zinc-800 dark:bg-zinc-900/40">
-            <h3 className="text-base font-semibold text-zinc-950 dark:text-zinc-50">
-                {t('empty_title')}
-            </h3>
-            <p className="mt-2 max-w-xs text-sm text-zinc-600 dark:text-zinc-400">
-                {t('empty_desc')}
-            </p>
-        </div>
-    );
-}
-
-type EmailBindingSuggestionStatus =
-    | 'pending'
-    | 'suggested'
-    | 'declined'
-    | 'accepted';
-
-const FEED_CARDS_BEFORE_EMAIL_BINDING_SUGGESTION = 20;
-
-function FeedReviewDeck({
-    cards,
-    isLoading,
-    isRefetching: _isRefetching,
-    isError,
-    errorMessage,
-    onRetry,
-    onReview,
-}: {
-    cards: FeedItem[];
-    isLoading: boolean;
-    isRefetching: boolean;
-    isError: boolean;
-    errorMessage: string | null;
-    onRetry: () => void;
-    onReview: (card: FeedItem, direction: SwipeDirection) => Promise<void>;
-}) {
-    const t = useTranslations('profile.feed');
-    const [selectedCard, setSelectedCard] = useState<FeedItem | null>(null);
-    const [isDialogOpen, setIsDialogOpen] = useState(false);
-    const [pendingCardId, setPendingCardId] = useState<string | null>(null);
-    const [isAnimating, setIsAnimating] = useState(false);
-    const isBusy = pendingCardId !== null;
-
-    const app = useAppContext();
-    const [swipeCount, setSwipeCount] = useState<number>(() => {
-        return parseInt(localStorage.getItem('feed-swipes') ?? '0', 10);
-    });
-    const isNeedSuggestEmail = useMemo(() => {
-        if (swipeCount > FEED_CARDS_BEFORE_EMAIL_BINDING_SUGGESTION) {
-            return false;
-        }
-        return !app.userDetails?.email;
-    }, [app.userDetails?.email, swipeCount]);
-
-    const [emailSuggestionStatus, setEmailSuggestionStatus] =
-        useState<EmailBindingSuggestionStatus>('pending');
-
-    const increaseSwipesCounter = () => {
-        const prevValue = parseInt(
-            localStorage.getItem('feed-swipes') ?? '0',
-            10,
-        );
-        const nextCount = prevValue + 1;
-        localStorage.setItem('feed-swipes', nextCount.toString());
-        setSwipeCount(nextCount);
-        return nextCount;
-    };
-
-    const handleCardClick = (card: FeedItem) => {
-        setSelectedCard(card);
-        setIsDialogOpen(true);
-    };
-
-    const handleReview = async (direction: SwipeDirection) => {
-        if (!selectedCard || isBusy) return;
-
-        setPendingCardId(getFeedItemKey(selectedCard));
-        setIsAnimating(true);
-
-        const updatedCount = increaseSwipesCounter();
-
-        if (
-            isNeedSuggestEmail &&
-            updatedCount === FEED_CARDS_BEFORE_EMAIL_BINDING_SUGGESTION
-        ) {
-            setEmailSuggestionStatus('suggested');
-        }
-
-        try {
-            await onReview(selectedCard, direction);
-
-            // Wait for fade out animation
-            await new Promise(resolve => setTimeout(resolve, 150));
-
-            // Find the current card index and show the next card
-            const currentIndex = cards.findIndex(
-                card => getFeedItemKey(card) === getFeedItemKey(selectedCard),
-            );
-            const nextIndex = currentIndex + 1;
-
-            if (nextIndex < cards.length) {
-                setSelectedCard(cards[nextIndex]);
-                setIsAnimating(false);
-            } else {
-                // No more cards, close the dialog
-                setIsDialogOpen(false);
-                setSelectedCard(null);
-            }
-        } finally {
-            setPendingCardId(null);
-        }
-    };
-
-    if (isLoading) {
-        return (
-            <div className="flex min-h-80 items-center justify-center rounded-xl border border-zinc-200 bg-white dark:border-zinc-800 dark:bg-zinc-950">
-                <Loader2 className="h-8 w-8 animate-spin text-zinc-400" />
-            </div>
-        );
-    }
-
-    if (isError) {
-        return (
-            <div className="flex min-h-80 flex-col items-center justify-center rounded-xl border border-red-200 bg-white px-6 text-center dark:border-red-900/60 dark:bg-zinc-950">
-                <Activity className="h-8 w-8 text-red-500" />
-                <p className="mt-4 text-sm text-zinc-700 dark:text-zinc-300">
-                    {errorMessage ?? t('queue_error')}
-                </p>
-                <Button className="mt-5 cursor-pointer" onClick={onRetry}>
-                    {t('retry')}
-                </Button>
-            </div>
-        );
-    }
-
-    if (cards.length === 0) {
-        return <FeedEmptyState />;
-    }
-
-    return (
-        <>
-            <div className="flex gap-2 overflow-x-auto pb-4">
-                {cards.map(card => {
-                    const badgeLabel = card.isRequest
-                        ? t('requests_badge')
-                        : card.isExtendedNetwork
-                          ? t('extended_network')
-                          : null;
-                    const avatarUrl = card.details.avatar
-                        ? createFileLink(card.details.avatar)
-                        : '';
-
-                    return (
-                        <div
-                            key={getFeedItemKey(card)}
-                            className="shrink-0 w-48 cursor-pointer"
-                            onClick={() => handleCardClick(card)}
-                        >
-                            <div className="flex flex-col items-center gap-3 bg-white dark:bg-zinc-950 hover:bg-zinc-50 hover:dark:bg-zinc-900 rounded-xl border border-zinc-200 dark:border-zinc-800 p-4 shadow-sm transition-colors min-h-58">
-                                <Avatar className="w-16 h-16 border border-zinc-200 dark:border-zinc-800">
-                                    <AvatarImage src={avatarUrl} />
-                                    <AvatarFallback className="text-sm font-semibold">
-                                        {card.details.nickname
-                                            .slice(0, 2)
-                                            .toUpperCase()}
-                                    </AvatarFallback>
-                                </Avatar>
-                                <div className="text-center min-w-0 flex-1">
-                                    <h3 className="truncate text-sm font-semibold text-zinc-950 dark:text-zinc-50">
-                                        {card.details.nickname}
-                                    </h3>
-                                    {badgeLabel && (
-                                        <Badge
-                                            variant="secondary"
-                                            className="mt-1 rounded-md text-xs"
-                                        >
-                                            {badgeLabel}
-                                        </Badge>
-                                    )}
-                                    <p className="mt-2 text-xs text-zinc-600 dark:text-zinc-400 overflow-hidden text-ellipsis">
-                                        {card.details.description
-                                            ? truncateString(
-                                                  card.details.description,
-                                                  60,
-                                              )
-                                            : t('no_description')}
-                                    </p>
-                                </div>
-                            </div>
-                        </div>
-                    );
-                })}
-            </div>
-
-            <StyledDialogWrapper
-                open={isDialogOpen}
-                onOpenChange={setIsDialogOpen}
-                contentClassName="top-0 bottom-0 lg:top-4 lg:bottom-4 lg:rounded-2xl max-w-md max-h-none -translate-x-1/2 -translate-y-0 bg-white dark:bg-zinc-950"
-            >
-                {selectedCard && (
-                    <FeedDialogContent
-                        selectedCard={selectedCard}
-                        isAnimating={isAnimating}
-                        closeDialog={() => {
-                            setIsDialogOpen(false);
-                            setSelectedCard(null);
-                        }}
-                        isBusy={isBusy}
-                        handleReview={direction => void handleReview(direction)}
-                    />
-                )}
-            </StyledDialogWrapper>
-            <SuggestEmailBindingDialog
-                status={emailSuggestionStatus}
-                setStatus={setEmailSuggestionStatus}
-            />
-            <EditProfileDialog
-                open={emailSuggestionStatus === 'accepted'}
-                setOpen={isOpen => {
-                    if (!isOpen) setEmailSuggestionStatus('declined');
-                }}
-            />
-        </>
-    );
-}
-
-interface FeedDialogProps {
-    selectedCard: FeedItem;
-    isAnimating: boolean;
-    isBusy: boolean;
-    closeDialog: () => void;
-    handleReview: (direction: SwipeDirection) => void;
-}
-
-function FeedDialogContent({
-    selectedCard,
-    isAnimating,
-    closeDialog,
-    isBusy,
-    handleReview,
-}: FeedDialogProps) {
-    const t = useTranslations('profile.feed');
-    const userAccessHashes = useUserAccessHashes();
-    const navigate = useNavigate();
-
-    async function routeToUser(friend: UserDetails) {
-        await userAccessHashes.service.save({
-            id: friend.id,
-            accessHash: friend.accessHash,
-        });
-        await navigate(`/user/${friend.id}`);
-    }
-
-    return (
-        <>
-            <Dialog.Title className="sr-only">
-                {selectedCard.details.nickname}
-            </Dialog.Title>
-
-            <div
-                className={`flex flex-col h-full transition-opacity duration-150 ${
-                    isAnimating ? 'opacity-0' : 'opacity-100'
-                }`}
-            >
-                <div className="absolute top-4 left-4 right-4 z-10 flex justify-between items-center gap-2">
-                    <div className="flex flex-column ms-4">
-                        {(selectedCard.isRequest ||
-                            selectedCard.isExtendedNetwork) && (
-                            <Badge
-                                variant="secondary"
-                                className="bg-secondary/50 -ms-4 me-5 backdrop-blur-md border rounded-md text-sm px-3 py-1"
-                            >
-                                {selectedCard.isRequest
-                                    ? t('requests_badge')
-                                    : selectedCard.isExtendedNetwork
-                                      ? t('extended_network')
-                                      : null}
-                            </Badge>
-                        )}
-                        {selectedCard.commonFriends.slice(0, 5).map(friend => {
-                            const avatar = friend.avatar;
-                            const fallback = getAvatarFallbackForNickname(
-                                friend.nickname,
-                            );
-                            return (
-                                <Avatar
-                                    className="w-10 h-10 -ms-4 border-2 border-white dark:border-zinc-800 cursor-pointer"
-                                    onClick={() => void routeToUser(friend)}
-                                >
-                                    <AvatarImage
-                                        src={
-                                            avatar
-                                                ? createFileLink(avatar)
-                                                : undefined
-                                        }
-                                    />
-                                    <AvatarFallback>
-                                        <span className="text-xl">
-                                            {fallback}
-                                        </span>
-                                    </AvatarFallback>
-                                </Avatar>
-                            );
-                        })}
-                    </div>
-                    <Button
-                        variant="ghost"
-                        size="sm"
-                        className="h-8 w-8 p-0 rounded-full bg-black/20 hover:bg-black/40 text-white"
-                        onClick={() => closeDialog()}
-                    >
-                        <X className="h-4 w-4" />
-                    </Button>
-                </div>
-
-                <div className="relative w-full aspect-square shrink-0 overflow-hidden">
-                    <Avatar className="w-full h-full rounded-none object-cover">
-                        <AvatarImage
-                            src={
-                                selectedCard.details.avatar
-                                    ? createFileLink(
-                                          selectedCard.details.avatar,
-                                      )
-                                    : undefined
-                            }
-                            className="object-cover w-full h-full"
-                        />
-                        <AvatarFallback className="text-6xl font-semibold w-full h-full flex items-center justify-center rounded-none bg-zinc-200 dark:bg-zinc-800">
-                            {selectedCard.details.nickname
-                                .slice(0, 2)
-                                .toUpperCase()}
-                        </AvatarFallback>
-                    </Avatar>
-
-                    <div className="absolute bottom-4 left-4 right-4 z-10 flex flex-wrap gap-2">
-                        {selectedCard.details.interests.map(interest => (
-                            <Badge
-                                key={interest}
-                                variant="secondary"
-                                className="px-2 py-1 dark:bg-zinc-800 dark:text-zinc-300 dark:hover:bg-zinc-700"
-                            >
-                                {interest}
-                            </Badge>
-                        ))}
-                    </div>
-                </div>
-
-                <div className="flex flex-col flex-1 shrink p-6 pb-12 relative">
-                    <h3 className="text-2xl font-semibold text-zinc-950 dark:text-zinc-50 mb-2">
-                        {selectedCard.details.nickname}
-                    </h3>
-
-                    <p className="text-sm leading-6 text-zinc-700 dark:text-zinc-300">
-                        {selectedCard.details.description ||
-                            t('no_description')}
-                    </p>
-                </div>
-
-                <div className="p-6 pt-0 relative">
-                    <div className="flex gap-4">
-                        <Button
-                            variant="outline"
-                            className="flex-1 h-12 cursor-pointer"
-                            disabled={isBusy}
-                            onClick={() => handleReview('left')}
-                        >
-                            <X className="h-5 w-5 mr-2" />
-                            {t('skip')}
-                        </Button>
-                        <Button
-                            className="flex-1 h-12 cursor-pointer"
-                            disabled={isBusy}
-                            onClick={() => handleReview('right')}
-                        >
-                            {selectedCard.isRequest ? (
-                                <Check className="h-5 w-5 mr-2" />
-                            ) : (
-                                <Heart className="h-5 w-5 mr-2" />
-                            )}
-                            {selectedCard.isRequest
-                                ? t('accept')
-                                : t('connect')}
-                        </Button>
-                    </div>
-                </div>
-            </div>
-        </>
-    );
-}
-
-interface SuggestEmailBindingDialogProps {
-    status: EmailBindingSuggestionStatus;
-    setStatus: (value: EmailBindingSuggestionStatus) => void;
-}
-
-function SuggestEmailBindingDialog({
-    status,
-    setStatus,
-}: SuggestEmailBindingDialogProps) {
-    const t = useTranslations('profile.feed');
-    return (
-        <StyledDialogWrapper
-            open={status === 'suggested'}
-            onOpenChange={() => setStatus('pending')}
-            contentClassName="-translate-y-1/2 w-fit max-w-sm max-h-none rounded-2xl p-6 bg-white dark:bg-zinc-950 shadow-lg"
-        >
-            <>
-                <h2 className="text-center">
-                    {t('email_binding.description')}
-                </h2>
-                <div className="w-full flex flex-row grow-1 gap-2 mt-4">
-                    <Button
-                        className="grow-1"
-                        onClick={() => setStatus('declined')}
-                    >
-                        {t('email_binding.decline')}
-                    </Button>
-                    <Button
-                        className="grow-2"
-                        onClick={() => setStatus('accepted')}
-                    >
-                        {t('email_binding.confirm')}
-                    </Button>
-                </div>
-            </>
-        </StyledDialogWrapper>
-    );
-}
+import {FriendsBlock} from '@/app/friends/friends-block';
+import {DiscoveryFeedBlock} from '@/app/discovery-feed-block';
+import {QrCodeCard} from '@/app/qrcode-card';
 
 function ProfileHeader({logOut}: {logOut: () => void}) {
     const t = useTranslations('profile');
@@ -557,142 +102,6 @@ function InterestsBlock({interests}: {interests: string[]}) {
                         {interest}
                     </Badge>
                 ))}
-            </div>
-        </div>
-    );
-}
-
-function DiscoveryFeedBlock() {
-    const t = useTranslations('profile.feed');
-    const backend = useBackend();
-
-    const feedQuery = useQuery({
-        queryKey: ['feedQueue'],
-        queryFn: () => backend.getFeedQueue(),
-    });
-
-    const [cards, setCards] = useState<FeedItem[]>([]);
-
-    useEffect(() => {
-        if (!feedQuery.data?.ok) {
-            return;
-        }
-
-        setCards(feedQuery.data.data.entries);
-    }, [feedQuery.data]);
-
-    const feedErrorMessage =
-        feedQuery.data && !feedQuery.data.ok
-            ? formatNetworkError(feedQuery.data.error)
-            : null;
-
-    const onReview = useCallback(
-        async (card: FeedItem, direction: SwipeDirection) => {
-            const request = {
-                userId: card.details.id,
-                userAccessHash: card.details.accessHash,
-            };
-
-            const result =
-                direction === 'right'
-                    ? await backend.sendFriendRequest(request)
-                    : await backend.declineFriendRequest(request);
-
-            if (!result.ok) {
-                toast.error(formatNetworkError(result.error));
-            }
-
-            setCards(current =>
-                current.filter(item => item.details.id !== card.details.id),
-            );
-        },
-        [backend],
-    );
-
-    return (
-        <section className="flex flex-col gap-5">
-            <div className="flex flex-col gap-2">
-                <div className="flex items-center gap-3">
-                    <div className="rounded-lg bg-secondary p-2 text-secondary-foreground">
-                        <Heart className="h-4 w-4" />
-                    </div>
-                    <div>
-                        <h2 className="text-base font-semibold text-zinc-950 dark:text-zinc-50">
-                            {t('title')}
-                        </h2>
-                        <p className="text-sm text-zinc-600 dark:text-zinc-400">
-                            {t('desc')}
-                        </p>
-                    </div>
-                </div>
-            </div>
-
-            <FeedReviewDeck
-                cards={cards}
-                isLoading={feedQuery.isLoading}
-                isRefetching={feedQuery.isRefetching}
-                isError={feedQuery.isError || Boolean(feedErrorMessage)}
-                errorMessage={feedErrorMessage}
-                onRetry={() => {
-                    void feedQuery.refetch();
-                }}
-                onReview={onReview}
-            />
-        </section>
-    );
-}
-
-function QrCodeCard({url}: {url: string | null}) {
-    const t = useTranslations('profile');
-    const queryClient = useQueryClient();
-    const backend = useBackend();
-
-    async function forceRefresh() {
-        await backend.friendsGenerateForce();
-        void queryClient.invalidateQueries({
-            queryKey: ['inviteToken'],
-        });
-    }
-
-    return (
-        <div className="md:w-1/4 md:h-fit md:mt-4 md:mr-8 flex flex-col items-center md:items-start p-4 md:rounded-xl md:border md:border-zinc-200 dark:md:border-zinc-800 md:bg-white dark:md:bg-zinc-900 text-sm">
-            <div className="flex flex-col gap-2 pl-2 pt-2 pr-2">
-                <div className="flex flex-row gap-2 items-center font-medium text-zinc-900 dark:text-zinc-100">
-                    <QrCodeIcon className="w-4 h-4" /> {t('qr.title')}
-                </div>
-                <p className="text-neutral-700 dark:text-zinc-400">
-                    {t('qr.desc')}
-                </p>
-            </div>
-            <div className="h-6 md:h-2" />
-            <div className="w-full flex flex-col items-center">
-                <div className="bg-white p-4 rounded-xl border border-zinc-200">
-                    {url ? (
-                        <QRCode value={url} className="w-32 h-32" />
-                    ) : (
-                        <Loader2 className="h-10 w-10 animate-spin text-zinc-400" />
-                    )}
-                </div>
-                <div className="h-2" />
-                <div className="flex gap-2">
-                    <Button
-                        variant="outline"
-                        className="flex-1 dark:bg-zinc-950 dark:hover:bg-zinc-800 cursor-pointer"
-                        onClick={() => {
-                            void navigator.clipboard.writeText(url ?? '');
-                            toast.success(t('qr.copied'));
-                        }}
-                    >
-                        <Copy className="w-4 h-4 mr-2" /> {t('qr.copy')}
-                    </Button>
-                    <Button
-                        variant="outline"
-                        className="flex-1 dark:bg-zinc-950 dark:hover:bg-zinc-800 cursor-pointer"
-                        onClick={() => void forceRefresh()}
-                    >
-                        <RotateCcw className="s-4" />
-                    </Button>
-                </div>
             </div>
         </div>
     );
@@ -829,16 +238,4 @@ export default function Home() {
             </div>
         </div>
     );
-}
-
-function getAvatarFallbackForNickname(
-    nickname: string | undefined,
-): string | undefined {
-    if (!nickname) return;
-    if (nickname.trim().length === 0) return;
-    const words = nickname.toUpperCase().split(' ');
-    return words
-        .slice(0, 2)
-        .map(word => word[0])
-        .join('');
 }
